fix: added card view for positions, employee levels[ref:#2061]

+++ apps/gauzy/src/app/pages/employee-levels/employee-level.component.ts
@@ -4,14 +4,17 @@ import {
 	Tag,
 	ComponentLayoutStyleEnum
 } from '@gauzy/models';
-import { NbToastrService } from '@nebular/theme';
+import { NbToastrService, NbDialogService } from '@nebular/theme';
 import { TranslateService } from '@ngx-translate/core';
 import { Subject } from 'rxjs';
 import { ComponentEnum } from '../../@core/constants/layout.constants';
-import { takeUntil } from 'rxjs/operators';
+import { takeUntil, first } from 'rxjs/operators';
 import { TranslationBaseComponent } from 'apps/gauzy/src/app/@shared/language-base/translation-base.component';
 import { EmployeeLevelService } from 'apps/gauzy/src/app/@core/services/employee-level.service';
 import { Store } from '../../@core/services/store.service';
+import { LocalDataSource } from 'ng2-smart-table';
+import { NotesWithTagsComponent } from '../../@shared/table-components/notes-with-tags/notes-with-tags.component';
+import { DeleteConfirmationComponent } from '../../@shared/user/forms/delete-confirmation/delete-confirmation.component';
 
 @Component({
 	selector: 'ga-employee-level',
@@ -29,16 +32,21 @@ export class EmployeeLevelComponent extends TranslationBaseComponent
 	employeeLevels: EmployeeLevelInput[] = [];
 	selectedEmployeeLevel: EmployeeLevelInput;
 	tags: Tag[] = [];
+	isGridEdit: boolean;
 	viewComponentName: ComponentEnum;
 	dataLayoutStyle = ComponentLayoutStyleEnum.TABLE;
+	settingsSmartTable: object;
+	smartTableSource = new LocalDataSource();
 
 	constructor(
 		private readonly employeeLevelService: EmployeeLevelService,
+		private dialogService: NbDialogService,
 		private readonly toastrService: NbToastrService,
 		private readonly store: Store,
 		readonly translateService: TranslateService
 	) {
 		super(translateService);
+		this.setView();
 	}
 
 	ngOnInit(): void {
@@ -48,6 +56,8 @@ export class EmployeeLevelComponent extends TranslationBaseComponent
 				if (organization) {
 					this.organizationId = organization.id;
 					this.loadEmployeeLevels();
+					this.loadSmartTable();
+					this._applyTranslationOnSmartTable();
 				}
 			});
 	}
@@ -60,6 +70,7 @@ export class EmployeeLevelComponent extends TranslationBaseComponent
 
 		if (items) {
 			this.employeeLevels = items;
+			this.smartTableSource.load(items);
 		}
 	}
 	setView() {
@@ -69,7 +80,24 @@ export class EmployeeLevelComponent extends TranslationBaseComponent
 			.pipe(takeUntil(this._ngDestroy$))
 			.subscribe((componentLayout) => {
 				this.dataLayoutStyle = componentLayout;
+				this.selectedEmployeeLevel =
+					this.dataLayoutStyle === 'CARDS_GRID'
+						? null
+						: this.selectedEmployeeLevel;
 			});
+	}
+	async loadSmartTable() {
+		this.settingsSmartTable = {
+			actions: false,
+			columns: {
+				level: {
+					title: this.getTranslation('ORGANIZATIONS_PAGE.LEVEL'),
+					type: 'custom',
+					class: 'align-row',
+					renderComponent: NotesWithTagsComponent
+				}
+			}
+		};
 	}
 	async addEmployeeLevel(level: string) {
 		if (level) {
@@ -88,9 +116,8 @@ export class EmployeeLevelComponent extends TranslationBaseComponent
 				),
 				this.getTranslation('TOASTR.TITLE.SUCCESS')
 			);
-
-			this.showAddCard = !this.showAddCard;
 			this.loadEmployeeLevels();
+			this.cancel();
 		} else {
 			this.toastrService.danger(
 				this.getTranslation(
@@ -110,30 +137,50 @@ export class EmployeeLevelComponent extends TranslationBaseComponent
 			tags: this.tags
 		};
 		await this.employeeLevelService.update(id, employeeLevel);
-
 		this.toastrService.primary(
 			this.getTranslation('TOASTR.MESSAGE.EMPLOYEE_LEVEL_UPDATE'),
 			this.getTranslation('TOASTR.TITLE.SUCCESS')
 		);
+
+		this.loadEmployeeLevels();
 		this.cancel();
-		this.loadEmployeeLevels();
-		this.tags = [];
 	}
-
+	edit(employeeLevel: EmployeeLevelInput) {
+		this.showAddCard = true;
+		this.isGridEdit = true;
+		this.selectedEmployeeLevel = employeeLevel;
+		this.tags = employeeLevel.tags;
+	}
+	save(name: string) {
+		console.log(this.isGridEdit);
+		if (this.isGridEdit) {
+			this.editEmployeeLevel(this.selectedEmployeeLevel.id, name);
+		} else {
+			this.addEmployeeLevel(name);
+		}
+	}
 	async removeEmployeeLevel(id: string, name: string) {
-		await this.employeeLevelService.delete(id);
-
-		this.toastrService.primary(
-			this.getTranslation(
-				'NOTES.ORGANIZATIONS.EDIT_ORGANIZATIONS_EMPLOYEE_LEVELS.REMOVE_EMPLOYEE_LEVEL',
-				{
-					name
+		const result = await this.dialogService
+			.open(DeleteConfirmationComponent, {
+				context: {
+					recordType: 'Employee level'
 				}
-			),
-			this.getTranslation('TOASTR.TITLE.SUCCESS')
-		);
-
-		this.loadEmployeeLevels();
+			})
+			.onClose.pipe(first())
+			.toPromise();
+		if (result) {
+			await this.employeeLevelService.delete(id);
+			this.toastrService.primary(
+				this.getTranslation(
+					'NOTES.ORGANIZATIONS.EDIT_ORGANIZATIONS_EMPLOYEE_LEVELS.REMOVE_EMPLOYEE_LEVEL',
+					{
+						name
+					}
+				),
+				this.getTranslation('TOASTR.TITLE.SUCCESS')
+			);
+			this.loadEmployeeLevels();
+		}
 	}
 
 	showEditCard(employeeLevel: EmployeeLevelInput) {
@@ -143,11 +190,19 @@ export class EmployeeLevelComponent extends TranslationBaseComponent
 	}
 
 	cancel() {
-		this.showEditDiv = !this.showEditDiv;
+		this.showEditDiv = false;
+		this.showAddCard = false;
 		this.selectedEmployeeLevel = null;
+		this.isGridEdit = false;
+		this.tags = [];
 	}
 
 	selectedTagsEvent(ev) {
 		this.tags = ev;
+	}
+	_applyTranslationOnSmartTable() {
+		this.translateService.onLangChange.subscribe(() => {
+			this.loadSmartTable();
+		});
 	}
 }
